Clean up ArmeriaServerOptions implmentation
